requirements.txt
sentence-transformers
faiss-cpu
numpy
scikit-learn
python-dotenv
openai
tiktoken